requirements.txt
facenet_pytorch
numpy
opencv_python
pandas==1.1.5
Pillow
retinaface==1.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
urllib3==1.26.6
